[#101] UI 베이스 대량 제작 및 골드 UI 개선

+++ Source/Project_LLL/Public/DataAsset/LLL_GoldComponentDataAsset.h
@@ -15,5 +15,5 @@ class PROJECT_LLL_API ULLL_GoldComponentDataAsset : public UDataAsset
 	GENERATED_BODY()
 public:
 	UPROPERTY(EditDefaultsOnly, Category = "Widget", DisplayName = "UI 노출 시간")
-	uint8 WidgetHideWaitTime;
+	float WidgetHideWaitTime;
 };
